Commiting changes before refactoring to try ordered by deep phase of dive.
