+++ DiveLog.API/Controllers/IBackgroundJobs.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Hangfire;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -7,6 +8,7 @@ namespace DiveLog.API.Controllers
 {
 	public interface IBackgroundJobs
 	{
+		[DisableConcurrentExecution(60)]
 		void DeriveDiveLogStatisics();
 		void SaveLogs(List<DAL.Models.LogEntry> logEntries);
 	}
